requirements.txt
streamlit
python-pptx
torch
transformers
PyMuPDF
requests